+++ app/models/post.rb
@@ -20,10 +20,11 @@ class Post < ActiveRecord::Base
   private
 
   def send_email
-    return if archived?
+    return true if archived?
     all_users = project.users_to_email - [user]
     all_users.each do |user_to_email|
       UserMailer.project_news(self, user_to_email).deliver_later if EMAILS_ENABLED
     end
+    true
   end
 end